requirements.txt
cassandra_driver==3.25.0
Flask==2.0.2
